+++ google_images_search/fetch_resize_save.py
@@ -1,6 +1,7 @@
 import os
-import shutil
+import curses
 import requests
+import threading
 from PIL import Image
 from resizeimage import resizeimage
 
@@ -10,10 +11,31 @@ from .google_api import GoogleCustomSearch
 class FetchResizeSave(object):
     """Class with resizing and downloading logic"""
 
-    def __init__(self, developer_key, custom_search_cx):
-        self._google_custom_search = GoogleCustomSearch(developer_key,
-                                                        custom_search_cx)
-        self._search_resut = []
+    def __init__(self, developer_key, custom_search_cx,
+                 progressbar_fn=None, progress=False):
+
+        # initialise google api
+        self._google_custom_search = GoogleCustomSearch(
+            developer_key, custom_search_cx, self)
+
+        self._search_result = list()
+
+        self._stdscr = None
+        self._progress = False
+        self._chunk_sizes = dict()
+        self._terminal_lines = dict()
+        self._download_progress = dict()
+        self._report_progress = progressbar_fn
+
+        if progressbar_fn:
+            # user nserted progressbar fn
+            self._progress = True
+        else:
+            if progress:
+                # initialise internal progressbar
+                self._progress = True
+                self._stdscr = curses.initscr()
+                self._report_progress = self.__report_progress
 
     def search(self, search_params, path_to_dir=False, width=None,
                height=None, cache_discovery=True):
@@ -27,24 +49,70 @@ class FetchResizeSave(object):
         :return: None
         """
 
-        for url in self._google_custom_search.search(search_params,
-                                                     cache_discovery):
+        i = 0
+        threads = list()
+        for url in self._google_custom_search.search(
+            search_params, cache_discovery
+        ):
+            # initialise image object
             image = GSImage(self)
             image.url = url
 
-            if path_to_dir:
-                image.download(path_to_dir)
-                if width and height:
-                    image.resize(width, height)
+            # set thread safe variables
+            self._download_progress[url] = 0
+            self._terminal_lines[url] = i
+            i += 2
 
-            self._search_resut.append(image)
+            # set thread with function and arguments
+            thread = threading.Thread(
+                target=self._download_and_resize,
+                args=(path_to_dir, image, width, height)
+            )
+
+            # start thread
+            thread.start()
+
+            # register thread
+            threads.append(thread)
+
+        # wait for all threads to end here
+        for thread in threads:
+            thread.join()
+
+        if self._progress:
+            if self._stdscr:
+                curses.endwin()
+
+    def set_chunk_size(self, url, content_size):
+        """Set images chunk size according to its size
+        :param url: image url
+        :param content_size: image size
+        :return: None
+        """
+
+        self._chunk_sizes[url] = int(int(content_size) / 100) + 1
+
+    def _download_and_resize(self, path_to_dir, image, width, height):
+        """Method used for threading
+        :param path_to_dir: path to download dir
+        :param image: image object
+        :param width: crop width
+        :param height: crop height
+        :return: None
+        """
+
+        if path_to_dir:
+            image.download(path_to_dir)
+            if width and height:
+                image.resize(width, height)
+        self._search_result.append(image)
 
     def results(self):
         """Returns objects of downloaded images
         :return: list
         """
 
-        return self._search_resut
+        return self._search_result
 
     def download(self, url, path_to_dir):
         """Downloads image from url to path dir
@@ -57,40 +125,37 @@ class FetchResizeSave(object):
         if not os.path.exists(path_to_dir):
             os.makedirs(path_to_dir)
 
-        raw_data = self.__class__.get_raw_data(url)
-
         raw_filename = url.split('/')[-1].split('?')[0]
         basename, ext = os.path.splitext(raw_filename)
         filename = "".join(x for x in basename if x.isalnum()) + ext
 
         path_to_image = os.path.join(path_to_dir, filename)
 
-        with open(path_to_image, 'wb') as f:
-            self.__class__.copy_to(raw_data, f)
+        with open(path_to_image, 'wb+') as f:
+            for chunk in self.get_raw_data(url):
+                f.write(chunk)
 
         return path_to_image
 
-    @staticmethod
-    def get_raw_data(url):
-        """Takes data from image url into a variable
+    def get_raw_data(self, url):
+        """Generator method for downloading images in chunks
         :param url: url to image
         :return: raw image data
         """
 
-        req = requests.get(url, stream=True)
-        req.raw.decode_content = True
-        return req.raw
+        with requests.get(url, stream=True) as req:
+            for chunk in req.iter_content(chunk_size=self._chunk_sizes[url]):
 
-    @staticmethod
-    def copy_to(raw_data, obj):
-        """
-        Copy raw image data to another object, preferably BytesIO
-        :param raw_data: raw image data
-        :param obj: BytesIO object
-        :return: None
-        """
+                # filter out keep-alive new chunks
+                if chunk:
 
-        shutil.copyfileobj(raw_data, obj)
+                    # report progress
+                    if self._progress:
+                        self._download_progress[url] += 1
+                        if self._download_progress[url] <= 100:
+                            self._report_progress(url, self._download_progress[url])
+
+                    yield chunk
 
     @staticmethod
     def resize(path_to_image, width, height):
@@ -106,6 +171,22 @@ class FetchResizeSave(object):
         img = resizeimage.resize_cover(img, [int(width), int(height)])
         img.save(path_to_image, img.format)
         fd_img.close()
+
+    def __report_progress(self, url, progress):
+        """Prints a progress bar in terminal
+        :param url:
+        :param progress:
+        :return:
+        """
+
+        self._stdscr.addstr(
+            self._terminal_lines[url], 0, "Downloading file: {0}".format(url)
+        )
+        self._stdscr.addstr(
+            self._terminal_lines[url] + 1, 0,
+            "Progress: [{1:100}] {0}%".format(progress, "#" * progress)
+        )
+        self._stdscr.refresh()
 
 
 class GSImage(object):
@@ -166,7 +247,7 @@ class GSImage(object):
         :return: raw data
         """
 
-        return self._fetch_resize_save.__class__.get_raw_data(self._url)
+        return b''.join(list(self._fetch_resize_save.get_raw_data(self._url)))
 
     def copy_to(self, obj, raw_data=None):
         """Copies raw image data to another object, preferably BytesIO
@@ -178,7 +259,7 @@ class GSImage(object):
         if not raw_data:
             raw_data = self.get_raw_data()
 
-        self._fetch_resize_save.__class__.copy_to(raw_data, obj)
+        obj.write(raw_data)
 
     def resize(self, width, height):
         """Resize the image