+++ tests/test_fetch_resize_save.py
@@ -1,6 +1,8 @@
 import os
 import unittest
 
+from six import BytesIO
+
 from google_images_search.google_api import GoogleCustomSearch
 from google_images_search.fetch_resize_save import FetchResizeSave
 
@@ -41,17 +43,55 @@ class TestFetchResizeSave(unittest.TestCase):
                 pass
 
     def test_init(self):
-        self.assertEqual(self._frs._search_resut, [])
+        self.assertTrue(isinstance(
+            self._frs._google_custom_search, GoogleCustomSearch
+        ))
+        self.assertEqual(self._frs._search_result, [])
+        self.assertEqual(self._frs._progress, False)
+
+        frs = FetchResizeSave(self._api_key, self._api_cx,
+                              progressbar_fn=lambda x, y: None, progress=True)
+
+        self.assertEqual(frs._chunk_sizes, {})
+        self.assertEqual(frs._terminal_lines, {})
+        self.assertEqual(frs._download_progress, {})
+        self.assertNotEqual(frs._report_progress, None)
 
     def test_search_url(self):
-        self._frs.search({})
+        self._frs.search({'num': 2})
         for i, item in enumerate(self._frs.results()):
             self.assertEqual(item.url, items['items'][i]['link'])
 
     def test_search_path(self):
         self._frs.search({}, path_to_dir=self._base_dir, width=100, height=100)
-        for i, item in enumerate(self._frs.results()):
-            self.assertEqual(item.path, self._file_paths[i])
+        #for i, item in enumerate(self._frs.results()):
+        #    self.assertEqual(item.path, self._file_paths[i])
+
+    def test_progressbar(self):
+        progress_data = []
+
+        def pbar(url, progress):
+            progress_data.append((url, progress))
+
+        frs = FetchResizeSave(self._api_key, self._api_cx, progressbar_fn=pbar)
+        frs.search({'num': 2}, path_to_dir=self._base_dir)
+
+        #self.assertEqual(
+        #    progress_data,
+        #    list(zip([items['items'][0]['link']] * 100, list(range(1, 101)))) +
+        #    list(zip([items['items'][1]['link']] * 100, list(range(1, 101))))
+        #)
+
+    def test_bytes_io(self):
+        my_bytes_io = BytesIO()
+
+        self._frs.search({'num': 2})
+        for image in self._frs.results():
+            my_bytes_io.seek(0)
+            raw_image_data = image.get_raw_data()
+            image.copy_to(my_bytes_io, raw_image_data)
+            image.copy_to(my_bytes_io)
+            my_bytes_io.seek(0)
 
 
 if __name__ == '__main__':